Study threshold.  Discovered problem with normalization.
Found results depend on initial sigma because training stopped before
convergence.  Fixed by improving initial sigma.

Found that by cheating and using the best threshold for each record
the error rate was still more than 10%

Changed the untrainable links and parameters.

Found that normalization eliminates the discovered peaks of heart rate
in x19.  But x19 has clear apnea oscillations.  After this commit, I
will revisit normalization in a branch.

+++ src/hmmds/applications/apnea/utilities.py
@@ -367,8 +367,8 @@ class HeartRate:
         if prominence is None:
             prominence = self.config.min_prominence
         s_f_hz = self.hr_sample_frequency.to('Hz').magnitude
-        distance_samples = distance.to('seconds').magnitude * s_f_hz
-        wlen_samples = wlen.to('seconds').magnitude * s_f_hz
+        distance_samples = int(distance.to('seconds').magnitude * s_f_hz)
+        wlen_samples = int(wlen.to('seconds').magnitude * s_f_hz)
 
         self.peaks, properties = scipy.signal.find_peaks(
             self.slow,
